Calculate minimal classpath manifest
The classpath manifest is now calculated by subtracting all
dependencies that are brought in transitively by subprojects.
This is necessary, because a constraint in the current project
can be triggered by a transitive dependency from a project dependency.
This would blow up the manifests.

Note that the classpath manifest includes now even less entries than
before since duplications are now removed - in cases where multiple
dependent projects bring in the same dependency.